Update halaman survey
