+++ lib/pages/survey.dart
@@ -1,103 +1,120 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:parkquest_parkir_gamifikasi/pages/form_survey.dart';
 
 class Survey extends StatelessWidget {
   const Survey({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          backgroundColor: Color(0xFFFEC827),
-          title: Text(
-            'Survey',
-            style: GoogleFonts.inter(
-              color: Colors.white,
-              fontWeight: FontWeight.w500,
-            ),
-          ),
-          leading: IconButton(
-            onPressed: () => Navigator.of(context).pop(),
-            icon: Icon(
-              CupertinoIcons.chevron_left,
-              color: Colors.white,
-            ),
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Color(0xFFFEC827),
+        title: Text(
+          'Survey',
+          style: GoogleFonts.inter(
+            color: Colors.white,
+            fontWeight: FontWeight.w500,
           ),
         ),
-        body: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: ListView(
-            children: [
-              InkWell(
-                onTap: () {
-                  Navigator.pushNamed(context, '/gedungutama');
-                },
-                child: Column(
-                  children: [
-                    buildSurveyCard('Gedung Utama', 'assets/img/login.png'),
-                    const Text('Go to Survey', style: TextStyle(fontSize: 12)),
-                  ],
-                ),
-              ),
-              InkWell(
-                onTap: () {
-                  Navigator.pushNamed(context, '/technopreneur');
-                },
-                child: Column(
-                  children: [
-                    buildSurveyCard('Technopreneur', 'assets/img/login.png'),
-                    const Text('Go to Survey', style: TextStyle(fontSize: 12)),
-                  ],
-                ),
-              ),
-              InkWell(
-                onTap: () {
-                  Navigator.pushNamed(context, '/tower');
-                },
-                child: Column(
-                  children: [
-                    buildSurveyCard('Tower A', 'assets/img/login.png'),
-                    const Text('Go to Survey', style: TextStyle(fontSize: 12)),
-                  ],
-                ),
-              ),
-            ],
+        leading: IconButton(
+          onPressed: () => Navigator.of(context).pop(),
+          icon: Icon(
+            CupertinoIcons.chevron_left,
+            color: Colors.white,
           ),
+        ),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.symmetric(
+          horizontal: 20,
+          vertical: 10,
+        ),
+        child: ListView(
+          children: [
+            // Gedung Utama
+            InkWell(
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => FormSurvey(),
+                  ),
+                );
+              },
+              child: buildSurveyCard('Gedung Utama', 'assets/img/login.png'),
+            ),
+            // Technopreneur
+            InkWell(
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => FormSurvey(),
+                  ),
+                );
+              },
+              child: buildSurveyCard('Technopreneur', 'assets/img/login.png'),
+            ),
+            // Tower A
+            InkWell(
+              onTap: () {
+                Navigator.push(
+                  context,
+                  MaterialPageRoute(
+                    builder: (context) => FormSurvey(),
+                  ),
+                );
+              },
+              child: buildSurveyCard('Tower A', 'assets/img/login.png'),
+            ),
+          ],
         ),
       ),
     );
   }
 
-  // Function to create each survey card
+  // Card
   Widget buildSurveyCard(String title, String imagePath) {
-    return Card(
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(20.0),
-      ),
-      elevation: 4,
-      child: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              title,
-              style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            title,
+            style: GoogleFonts.inter(
+              fontSize: 18,
+              fontWeight: FontWeight.bold,
             ),
-            SizedBox(height: 8),
-            ClipRRect(
+          ),
+          Card(
+            color: Colors.white,
+            shape: RoundedRectangleBorder(
               borderRadius: BorderRadius.circular(20.0),
-              child: Image.asset(
-                imagePath,
-                height: 150,
-                width: double.infinity,
-                fit: BoxFit.cover,
+            ),
+            elevation: 3,
+            child: Padding(
+              padding: const EdgeInsets.all(16),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  SizedBox(height: 8),
+                  ClipRRect(
+                    borderRadius: BorderRadius.circular(20.0),
+                    child: Image.asset(
+                      imagePath,
+                      height: 150,
+                      width: double.infinity,
+                      fit: BoxFit.cover,
+                    ),
+                  ),
+                ],
               ),
             ),
-          ],
-        ),
+          ),
+        ],
       ),
     );
   }
