+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   image_picker: ^1.1.2
   flutter_dotenv: ^5.2.1
   flutter_rating_bar: ^4.0.1
+  video_player: ^2.9.2
 
 dev_dependencies:
   flutter_test:
